+++ app/Http/Controllers/LibrarianController.php
@@ -13,6 +13,8 @@ class LibrarianController extends Controller
 {
     public function addBook(Request $req)
     {
+        if (Auth::user()->role != 'librarian')
+            return $this->message('Non Authorize', 1001);
         $name = $req->get('name');
         $author = $req->get('author');
         $description = $req->get('description');
@@ -27,32 +29,34 @@ class LibrarianController extends Controller
                 'year' => $year
             ]
         );
-        return $this->message('Ok',200);
+        return $this->message('Ok', 200);
     }
 
     public function deleteBook(Request $req)
     {
+        if (Auth::user()->role != 'librarian')
+            return $this->message('Non Authorize', 1001);
         $book_id = $req->get('book_id');
         Book::find($book_id)->delete();
-        return $this->message('Ok',200);
+        return $this->message('Ok', 200);
     }
 
-    public function sendBook(Request $req)
-    {
-        $librarian_id = Auth::id();
-        $user_book_id = $req->get('user_book_id');
-        UserBook::where('id', '=', $user_book_id)->update(['librarian_id' => $librarian_id]);
-        return $this->message('Ok',200);
-    }
+    // public function sendBook(Request $req)
+    // {
+    //     $librarian_id = Auth::id();
+    //     $user_book_id = $req->get('user_book_id');
+    //     UserBook::where('id', '=', $user_book_id)->update(['librarian_id' => $librarian_id]);
+    //     return $this->message('Ok',200);
+    // }
 
-    public function receiveBook(Request $req)
-    {
-        $book_id = $req->get('book_id');
-        UserBook::where(
-            'book_id',
-            '=',
-            $book_id
-        )->delete();
-        return $this->message('Ok',200);
-    }
+    // public function receiveBook(Request $req)
+    // {
+    //     $book_id = $req->get('book_id');
+    //     UserBook::where(
+    //         'book_id',
+    //         '=',
+    //         $book_id
+    //     )->delete();
+    //     return $this->message('Ok',200);
+    // }
 }
